fix(clouds): 레이 시작점 지터 복구 — 덜컥거림의 본체였다
패치 2(half -> float)로 줄긴 했지만 남아 있었다. 원인이 정밀도 하나가
아니라 레이마칭 표본 분포에도 있었다.

1. 지터가 사실상 안 걸려 있었다 (본체)

   HDRP 원본 : currentDistance = integrationNoise * stepS;
   이 패키지 : currentDistance = integrationNoise;      // * stepS 누락

   integrationNoise 는 0~1 이다. 거리(미터)에 그대로 넣으면 지터 폭이
   최대 1m 인데 이 씬의 stepS 는 41.67m 다. 스텝 간격의 2.4% 라
   사실상 안 걸린 것과 같았다. 시작점이 안 흩어지면 모든 픽셀이 같은
   평면에서 밀도를 떠서 계단과 띠가 남는다.

   지터 폭 1.00m -> 41.67m (stepS 대비 2.4% -> 100%)

2. 노이즈가 프레임마다 안 바뀌었다

   float time = unity_DeltaTime.y * _Time.y + _Seed;
   _Seed += 1.0;   // 효과 없음

   _Seed 는 CBUFFER 상수라 셰이더에서 쓴 값이 밖으로 안 나간다.
   unity_DeltaTime.y 는 1/dt 라 프레임률이 안정되면 매 프레임 거의
   같은 값이다. 결국 같은 패턴이 반복되고, 재투영이 같은 오차를 계속
   누적한다. 플레이 모드에서 더 도드라지던 직접 원인이다.

3. 화이트노이즈 -> R2 저불일치 수열

   GenerateHashedRandomFloat 은 Jenkins 해시, 즉 화이트노이즈다
   (업스트림 주석의 "blue noise" 는 사실이 아니다). 이웃 픽셀 값이
   독립이라 우연히 뭉친 구역이 저주파 얼룩으로 보인다.

   R2 수열은 이미 찍은 점에서 최대한 먼 곳을 차례로 고른다. 공간
   분포와 프레임 간 회전을 한 식에서 같이 처리한다.

     공간 균일성      0.0335 -> 0.0108  (3.1배)
     시간 수렴 8프레임  0.0820 -> 0.0330  (2.5배)
     시간 수렴 32프레임 0.0408 -> 0.0108  (3.8배)

   누적이 쌓일수록 격차가 벌어진다. 이 씬은 _AccumulationFactor 0.95 로
   길게 누적하므로 그 성질이 그대로 이득이다.

   비용은 오히려 준다. 곱셈 2 + 덧셈 2 + frac 1 로 Jenkins 해시보다
   싸고, 블루노이즈 텍스처와 달리 텍스처 샘플이 0 이다.

4. relativeStepSize 제거

   lerp(integrationNoise, 1.0, saturate(currentIndex)) 인데 currentIndex 가
   int 라 saturate 가 0/1 로만 떨어진다. 첫 스텝만 noise 배이고 2번째부터
   전부 1.0 이다 — 스텝을 흩는 게 아니라 첫 칸만 좁히는 동작이었다.
   게다가 1번 수정으로 시작점을 이미 noise*stepS 밀어놨으므로 같은
   난수를 두 번 먹어 지터 효과를 깎는다.

   HDRP 원본도 stepS 로 균일 전진한다. 빼는 쪽이 원본과 일치하고 더 싸다.

5. _NumPrimarySteps, _MaxStepSize half -> float

   stepS 를 정하는 두 값이다. stepS 가 표본 간격이자 지터 폭이라
   정밀도가 곧 품질이다. 다만 _MaxStepSize(250)는 현재 설정에서
   min() 에 안 뽑히므로 실재 오차는 _NumPrimarySteps 쪽 0.0104m 뿐이다.
   비용 0 이고 설정 바꾸면 즉시 물리는 자리라 올려 뒀다.

   Defs.hlsl 의 나머지 half 27개는 전수 점검했고 그대로 둔다. 0~1 계수,
   색, SH 계수라 half 로 충분하고 올리면 레지스터만 낭비다.

아직 에디터에서 컴파일·화면 확인을 못 했다. MCP 브리지가 꺼져 있어
재임포트를 걸 수단이 없었다. 수치 검증만 거친 상태다.

+++ unity/DRON/Assets/VolumetricClouds/VolumetricCloudsDefs.hlsl
@@ -3,9 +3,23 @@
 
 CBUFFER_START(UnityPerMaterial)
 float _Seed;
-half _NumPrimarySteps;
+// [여의도63 수정] half -> float. 이 둘이 레이마칭 스텝 간격을 정한다.
+//
+//   VolumetricClouds.hlsl:61 에서
+//       float stepS = min(totalDistance / _NumPrimarySteps, _MaxStepSize);
+//   로 stepS 가 나오고, 이게 표본 간격이자 시작점 지터의 폭이다.
+//
+//   _MaxStepSize 는 보통 수백~수천 단위라 half 의 표현 간격이 벌어지는
+//   구간에 들어간다. 거기서 값이 계단으로 떨어지면 스텝 간격이 프레임마다
+//   튀고, 밀도 표본 위치가 통째로 흔들려 구름이 덜컥거린다.
+//   위치 누적값을 float 로 올린 것(패치 2)과 같은 병이다.
+//
+//   _NumPrimarySteps 는 나눗셈의 분모라 같이 올린다. 48 정도는 half 로도
+//   정확하지만, stepS 를 float 로 계산시키려면 분모도 float 여야 중간에서
+//   정밀도가 깎이지 않는다.
+float _NumPrimarySteps;
 half _NumLightSteps;
-half _MaxStepSize;
+float _MaxStepSize;
 float _HighestCloudAltitude;
 float _LowestCloudAltitude;
 // [여의도63 수정] half -> float. 구름 흐름 오프셋이다.
